nearly finished

+++ philo/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gforns-s <gforns-s@student.42.fr>          +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 13:30:39 by gforns-s          #+#    #+#             */
-/*   Updated: 2024/04/24 21:06:56 by gforns-s         ###   ########.fr       */
+/*   Updated: 2024/04/25 02:25:43 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,10 +39,10 @@ typedef struct s_philo
 	int				meals;
 	pthread_t		tid;
 	struct s_data	*data;
-	uint64_t		lst_meal;
-	uint64_t		t_death;
-	uint64_t		t_eat;
-	uint64_t		t_sleep;
+	unsigned long long		lst_meal;
+	unsigned long long		t_death;
+	unsigned long long		t_eat;
+	unsigned long long		t_sleep;
 	int				nb_meal;
 	pthread_mutex_t	lock;
 	pthread_mutex_t	*r_fork;
@@ -52,13 +52,13 @@ typedef struct s_philo
 typedef struct s_data
 {
 	int				nb_philo;
-	uint64_t		t_death;
-	uint64_t		t_eat;
-	uint64_t		t_sleep;
+	unsigned long long		t_death;
+	unsigned long long		t_eat;
+	unsigned long long		t_sleep;
 	int				nb_meal;
 	int				finished;
 	pthread_mutex_t	*fork;	
-	uint64_t		t0;
+	unsigned long long		t0;
 	pthread_mutex_t	print;
 	pthread_mutex_t	lock;
 	pthread_mutex_t	dead;
@@ -71,19 +71,19 @@ int			ft_atoi(const char *str);
 int			init_philos(t_data *data);
 int			create_philos(t_data *data);
 void		*routine(void *aux);
-uint64_t	get_time_ms(void);
-uint64_t	get_curr_time_f(t_data *data);
-uint64_t	eat_time(t_philo *philo);
-uint64_t	sleep_time(t_philo *philo);
-uint64_t	death_time(t_philo *philo);
-uint64_t	init_time(t_data *data);
-uint64_t	lst_meal_time(t_philo *philo);
+unsigned long long	get_time_ms(void);
+unsigned long long	get_curr_time_f(t_data *data);
+unsigned long long	eat_time(t_philo *philo);
+unsigned long long	sleep_time(t_philo *philo);
+unsigned long long	death_time(t_philo *philo);
+unsigned long long	init_time(t_data *data);
+unsigned long long	lst_meal_time(t_philo *philo);
 int			ft_is_dead(t_data *data);
 int			check_philo_status(t_philo *philo);
-void		ft_usleep(uint64_t time);
+void		ft_usleep(unsigned long long time);
 void		p_meals(t_philo *philo);
-void		ft_print_p(t_philo *philo, uint64_t t, int p, char *s);
-uint64_t	get_curr_time(t_data *data);
+void		ft_print_p(t_philo *philo, unsigned long long t, int p, char *s);
+unsigned long long	get_curr_time(t_data *data);
 int			create_one_philo(t_data *data);
 int			ft_strcmp(const char *s1, const char *s2);
 int			get_ph_meals(t_philo *philo);
